Refactored module name and unit test
