+++ __init__.py
@@ -1,7 +1,7 @@
 # __init__.py
 
 # Import classes or functions from app.py
-from .app import substitute_text  # Assuming main_function is a function in app.py
+from .substitute_text import substitute_text  # Assuming main_function is a function in app.py
 
 # Import mappings from mappings.py
 from .mappings import mappings
