+++ substitute_text.py
